codegen.rs に移管
